backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
python-dotenv==1.0.1
langfuse==2.21.0
openai==1.12.0
anthropic==0.18.0
google-generativeai==0.3.2
langchain==0.1.6
langchain-openai==0.0.5


groq>=0.37.0
sentry-sdk[fastapi]>=2.47.0
